* Now creating '/var/lib/dbus/machine-id' as symlink to '/etc/machine-id', if '/var/lib/dbus/' exists, * added backup creation and sanity checks.

+++ random_machine_id.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-VERSION=20230728.01
+VERSION=20230728.02
+
+set -e # Abort on errors, so that a failing cd does not mess up elsewhere.
 
 while [ "$#" -ge 1 ]; do
   case "$1" in
@@ -19,10 +21,22 @@ done
 
 ### The idea for the following is taken from https://forum.artixlinux.org/index.php/topic,1547.msg10599.html#msg10599:
 
-rm -f /etc/machine-id
+if [ -e /etc/machine-id ]; then
+  if [ -e /etc/machine-id.prev ]; then
+    rm -f /etc/machine-id.prev
+  fi
+  cp /etc/machine-id /etc/machine-id.prev # Create a backup
+  rm -f /etc/machine-id
+fi
 dbus-uuidgen --ensure=/etc/machine-id
 
 if [ -d /var/lib/dbus ]; then
-  rm -f /var/lib/dbus/machine-id
-  dbus-uuidgen --ensure=/var/lib/dbus/machine-id
+  if [ -e /var/lib/dbus/machine-id ]; then
+    if [ -e /var/lib/dbus/machine-id.prev ]; then
+      rm -f /var/lib/dbus/machine-id.prev
+    fi
+    cp /var/lib/dbus/machine-id /var/lib/dbus/machine-id.prev # Create a backup
+    rm -f /var/lib/dbus/machine-id
+  fi
+  ln -s /etc/machine-id /var/lib/dbus/machine-id
 fi
